get_info_tool/requirements.txt
urllib3
certifi
builtwith
python-whois


